refactor TodoList into separate component
